refactor: reorganize modules and build logic
- Move Use method from ModuleF to Project
- Simplify Builder by removing newRepo injection
- Fix source directory and build output paths
- Update tests and fixtures accordingly
