+++ formula/classfile.go
@@ -2,13 +2,10 @@ package formula
 
 import (
 	"os"
-	"path/filepath"
 	"runtime"
 	"sort"
 	"strings"
 
-	"github.com/goplus/llar/internal/env"
-	"github.com/goplus/llar/pkgs/mod/module"
 	"github.com/goplus/llar/pkgs/mod/versions"
 	"github.com/qiniu/x/gsh"
 )
@@ -182,61 +179,6 @@ func appendFlag(key, flag string) {
 		os.Setenv(key, flag)
 	} else {
 		os.Setenv(key, strings.TrimSpace(current+" "+flag))
-	}
-}
-
-// Use configures the build environment to use the specified module.
-func (p *ModuleF) Use(mod module.Version, matrix Matrix) {
-	if len(matrix.Require) == 0 {
-		return
-	}
-	formulaDir, err := env.FormulaDir()
-	if err != nil {
-		return
-	}
-	espcapedId, err := module.EscapeID(mod.ID)
-	if err != nil {
-		return
-	}
-	buildDir := filepath.Join(formulaDir, espcapedId, "build", mod.Version, matrix.String())
-
-	includeDir := filepath.Join(buildDir, "include")
-	libDir := filepath.Join(buildDir, "lib")
-	pkgconfigDir := filepath.Join(buildDir, "lib", "pkgconfig")
-
-	// PKG_CONFIG_PATH - pkg-config path (all platforms)
-	if _, err := os.Stat(pkgconfigDir); err == nil {
-		prependEnv("PKG_CONFIG_PATH", pkgconfigDir)
-	}
-
-	// CMAKE paths (all platforms)
-	if _, err := os.Stat(buildDir); err == nil {
-		prependEnv("CMAKE_PREFIX_PATH", buildDir)
-	}
-	if _, err := os.Stat(includeDir); err == nil {
-		prependEnv("CMAKE_INCLUDE_PATH", includeDir)
-	}
-	if _, err := os.Stat(libDir); err == nil {
-		prependEnv("CMAKE_LIBRARY_PATH", libDir)
-	}
-
-	// Platform-specific settings
-	if runtime.GOOS == "windows" {
-		// Windows MSVC environment variables
-		if _, err := os.Stat(includeDir); err == nil {
-			prependEnv("INCLUDE", includeDir)
-		}
-		if _, err := os.Stat(libDir); err == nil {
-			prependEnv("LIB", libDir)
-		}
-	} else {
-		// Unix (Linux/macOS) - Autotools/GCC style flags
-		if _, err := os.Stat(includeDir); err == nil {
-			appendFlag("CPPFLAGS", "-I"+includeDir)
-		}
-		if _, err := os.Stat(libDir); err == nil {
-			appendFlag("LDFLAGS", "-L"+libDir)
-		}
 	}
 }
